fixing 5 minute cache on chapter upload/purchase, fixing transaction logic

+++ Application/DTOs/Chapter/UnlockChapterResponseDto.cs
@@ -6,11 +6,11 @@ using System.Threading.Tasks;
 
 namespace Application.DTOs.Chapter
 {
-  public class UnlockChapterResponseDto
-  {
-    public int ChapterId { get; set; }
-    public long CoinsSpent { get; set; }
-    public decimal NewCoinBalance { get; set; }
-    public string Message { get; set; } = string.Empty;
-  }
+    public class UnlockChapterResponseDto
+    {
+        public int ChapterId { get; set; }
+        public decimal CoinsSpent { get; set; }
+        public decimal NewCoinBalance { get; set; }
+        public string Message { get; set; } = string.Empty;
+    }
 }
